訓練的指令敆做一个，較好管理
